refactor: isolate Android snapshot timeout evidence
